pyproject.toml
[project]
name = "lgpd-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "faker>=40.11.0",
    "fastapi[standard]>=0.135.1",
    "httpx>=0.28.1",
    "langchain-text-splitters>=0.3.3",
    "lxml>=6.0.2",
    "openai>=2.28.0",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "sqlalchemy>=2.0.48",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.6",
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py313"
src = ["."]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # flake8 errors
    "I",    # import sorting
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "S",    # flake8-strict
]
ignore = [
    "B008", # Do not perform function call `Depends` in argument defaults (standard FastAPI pattern)
]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
